refactor: handleSubmitのリフトアップ
