+++ src/Article.tsx
@@ -1,5 +1,6 @@
 import { h } from 'preact';
 import { useState, useEffect, useRef, Ref } from 'preact/hooks';
+import { v4 as uuidv4 } from 'uuid';
 import { Editor } from './Editor';
 import { useSubmit } from './useSubmit';
 import { articlable, emptyArticle } from './types';
@@ -21,10 +22,22 @@ export const Article = (props: {
       .then(article => setArticle(article));
   }, []);
 
+  const handleSubmit = () => {
+    submit({
+      text: editorRef.current.value(),
+      tags: ['dummyTag1', 'dummyTag2'],
+      uuid: props.isNew ? uuidv4().replace(/-/g, '') : article.uuid,
+      starred: false,
+    }, props.isNew);
+    props.setModified();
+    if (props.isNew) localStorage.setItem('smde_new', '');
+  };
+
   return (
     <Editor
       article={article}
       isNew={props.isNew}
+      handleSubmit={handleSubmit}
       submit={submit}
       isSubmitting={isSubmitting}
       setModified={props.setModified}
